pyproject.toml
[project]
name = "graph-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agent-framework>=1.0.0b260107",
    "agent-framework-devui>=1.0.0b260107",
    "azure-eventhub>=5.15.1",
    "azure-eventhub-checkpointstoreblob>=1.2.0",
    "azure-eventhub-checkpointstoreblob-aio>=1.2.0",
    "azure-identity>=1.25.1",
    "azure-storage-blob>=12.28.0",
    "gremlinpython>=3.8.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
